LayoutTests/imported/w3c/web-platform-tests/html/links/stylesheet/quirk-origin-check-recursive-import-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'data:/,@import%20url('x/');' because non CSS MIME types are not allowed for cross-origin stylesheets.

PASS Origin check for stylesheet with non-CSS MIME type quirk: recursive @import
